feat(client): HASHI-108/HASHI-109 홈 페이지/검색 페이지 공통 API 구조 변경
