+++ apps/client/src/pages/search/queries/searchRestaurantQueryKeys.ts
@@ -1,12 +1,7 @@
-import type { GetRestaurantsParams } from '@/pages/search/api/getRestaurants'
 import type { GetSearchKeywordRecommendationsParams } from '@/pages/search/api/getSearchKeywordRecommendations'
 
 export const searchRestaurantQueryKeys = {
   all: ['searchRestaurants'] as const,
-  infiniteLists: () =>
-    [...searchRestaurantQueryKeys.all, 'infiniteList'] as const,
-  infiniteList: (params: GetRestaurantsParams) =>
-    [...searchRestaurantQueryKeys.infiniteLists(), params] as const,
   keywordRecommendations: (
     params: GetSearchKeywordRecommendationsParams = {},
   ) =>
@@ -15,7 +10,4 @@ export const searchRestaurantQueryKeys = {
       'keywordRecommendations',
       params,
     ] as const,
-  lists: () => [...searchRestaurantQueryKeys.all, 'list'] as const,
-  list: (params: GetRestaurantsParams) =>
-    [...searchRestaurantQueryKeys.lists(), params] as const,
 }
